pyproject.toml
[tool.poetry]
name = "state-fin-api"
version = "0.1.0"
description = ""
authors = ["Poff Poffenberger <poffdeluxe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.11.8"
fastapi = "^0.61.1"
elasticsearch = {extras = ["async"], version = "^7.9.1"}
python-dotenv = "^0.14.0"
pyhumps = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
